Update daily build number to follow SemVer v1 rules.
Per SemVer v1 the pre-release version number is sorted by lexicographic
order.  To ensure that build-10 is greater than build-2 we must pad the
build number with the appropriate number of preceeding zeroes.  In this
cas the value is an unsigned 16-bit value in base 10 making a total of
five characters (e.g. build-00010 > build-00002).

Renamed GetNextDailyBuildNumber to GetNextBuildNumber and updated all
references.

Update build tools package dependency to latest build.
