+++ dir.props
@@ -1,7 +1,7 @@
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <!-- Build Tools Version -->
   <PropertyGroup>
-    <BuildToolsVersion>1.0.25-prerelease-5</BuildToolsVersion>
+    <BuildToolsVersion>1.0.25-prerelease-00005</BuildToolsVersion>
   </PropertyGroup>
 
   <!-- Common repo directories -->
