+++ src/.nuget/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.DotNet.BuildTools" version="1.0.25-prerelease-5" />
+  <package id="Microsoft.DotNet.BuildTools" version="1.0.25-prerelease-00005" />
   <package id="xunit.runners" version="2.0.0-beta5-build2785" />
 </packages>
