+++ src/BuildNumber.props
@@ -1,7 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
-    <!-- the value of DailyBuildNumber must be numeric -->
-    <DailyBuildNumber>5</DailyBuildNumber>
+    <!--
+      the value of BuildNumber must be numeric and contain enough
+      leading zeroes to satisfy a 16-bit integer in base-10 format.  this
+      is needed because SemVer v1 performs a lexical comparison of the
+      prerelease version number and without the leading zeroes foo-20 is
+      smaller than foo-4.
+    -->
+    <BuildNumber>00005</BuildNumber>
   </PropertyGroup>
 </Project>
